TKWAR-18: Implementing code quality checks. NOTE: Testing some plugins and cleaning the code.

+++ src/main/scala/com/colofabrix/scala/simulation/integration/TankEvaluator.scala
@@ -53,7 +53,7 @@ object TankEvaluator {
    */
   def fitness( t: Tank ): Double = {
     val ( w, ts ) = ( t.world, t.world.tanks )
-    max( t.points, 0.0 )
+    max( t.points.toDouble, 0.0 )
   }
 
   /**
@@ -63,7 +63,7 @@ object TankEvaluator {
    * @return The highest possible value
    */
   def higherFitness( world: World ): Double = {
-    world.tanks.filter( !_.isDead ).maxBy( _.points ).points
+    world.tanks.filter( !_.isDead ).maxBy( _.points ).points.toDouble
   }
 
 }
